refactor: move converter to Inserter
